requirements.txt
# Django
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# PDF Processing
PyMuPDF>=1.23.0
Pillow>=10.0.0
reportlab>=4.0.0

# Utilities
python-dotenv>=1.0.0
